feat: upgrade to esm module
This change migrates the commonjs code to esm.  In addition to the index
refactor, the `package.json` type has been changed to `"module"` and
an `exports` entry was added for modern bundler tooling.

The `xo` dependency was way behind, and after updating to the latest
version, it revealed a ton of eslint errors, that I also addressed.